+++ test/phone_form_field_test.dart
@@ -1,8 +1,13 @@
 import 'package:circle_flags/circle_flags.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:phone_form_field/phone_form_field.dart';
 import 'package:phone_form_field/src/models/phone_number_input.dart';
+import 'package:phone_form_field/src/validator/phone_validator.dart';
+
+typedef PhoneNumberInputValidatorDelegate = PhoneNumberInputValidator Function(
+    BuildContext context);
 
 void main() {
   group('PhoneFormField', () {
@@ -13,19 +18,26 @@ void main() {
       PhoneController? controller,
       bool showFlagInInput = true,
       String defaultCountry = 'US',
-      PhoneNumberType? phoneNumberType,
+      PhoneNumberInputValidatorDelegate? validatorDelegate,
     }) =>
         MaterialApp(
+          localizationsDelegates: [
+            GlobalMaterialLocalizations.delegate,
+            PhoneFieldLocalization.delegate,
+          ],
+          supportedLocales: [Locale('en')],
           home: Scaffold(
-            body: PhoneFormField(
-              initialValue: initialValue,
-              onChanged: onChanged,
-              onSaved: onSaved,
-              showFlagInInput: showFlagInInput,
-              controller: controller,
-              defaultCountry: defaultCountry,
-              phoneNumberType: phoneNumberType,
-            ),
+            body: Builder(builder: (context) {
+              return PhoneFormField(
+                initialValue: initialValue,
+                onChanged: onChanged,
+                onSaved: onSaved,
+                showFlagInInput: showFlagInInput,
+                controller: controller,
+                defaultCountry: defaultCountry,
+                validator: validatorDelegate?.call(context),
+              );
+            }),
           ),
         );
 
@@ -175,15 +187,22 @@ void main() {
       PhoneNumber? phoneNumber = PhoneParser().parseWithIsoCode('BE', '');
       // valid fixed line
       await tester.pumpWidget(getWidget(
-          initialValue: phoneNumber,
-          phoneNumberType: PhoneNumberType.fixedLine));
+        initialValue: phoneNumber,
+        validatorDelegate: (context) =>
+            PhoneValidator.invalidFixedLine(context),
+      ));
       final foundTextField = find.byType(TextFormField);
       await tester.enterText(foundTextField, '77777777');
       await tester.pumpAndSettle();
       expect(find.text('Invalid'), findsNothing);
       // invalid mobile
       await tester.pumpWidget(getWidget(
-          initialValue: phoneNumber, phoneNumberType: PhoneNumberType.mobile));
+        initialValue: phoneNumber,
+        validatorDelegate: (context) => PhoneValidator.invalidMobile(
+          context,
+          errorText: 'Invalid phone number',
+        ),
+      ));
       final foundTextField2 = find.byType(TextFormField);
       await tester.pumpAndSettle();
       await tester.enterText(foundTextField2, '77777777');
@@ -192,7 +211,12 @@ void main() {
 
       // valid mobile
       await tester.pumpWidget(getWidget(
-          initialValue: phoneNumber, phoneNumberType: PhoneNumberType.mobile));
+        initialValue: phoneNumber,
+        validatorDelegate: (context) => PhoneValidator.invalidMobile(
+          context,
+          errorText: 'Invalid phone number',
+        ),
+      ));
       final foundTextField3 = find.byType(TextFormField);
       await tester.enterText(foundTextField3, '477668899');
       await tester.pumpAndSettle();